Refactor driver to only depend on imxrt-ral
The driver depended on teensy4-bsp as a convenience. But, we only need
imxrt-ral to implement the driver. With this commit, we can more easily
explore this driver on another, non-Teensy system.

I broke the imxrt-hal dependency with a FnMut that, by convention,
implements the blocking delay needed for descriptor printing. See the
inline comments for more information. Everything still builds, but know
that I don't have the hardware to test this change.

The test for the Teensy 4.1 is now an example. This lets us remove more
dependencies from the package.

cortex-m and embedded-hal appear to be unused, so I dropped them.